+++ core-ng/src/main/java/core/framework/impl/web/management/HealthCheckController.java
@@ -1,5 +1,6 @@
 package core.framework.impl.web.management;
 
+import core.framework.api.http.HTTPStatus;
 import core.framework.api.web.Controller;
 import core.framework.api.web.Request;
 import core.framework.api.web.Response;
@@ -10,6 +11,6 @@ import core.framework.api.web.Response;
 public class HealthCheckController implements Controller {
     @Override
     public Response execute(Request request) throws Exception {
-        return Response.text("", null);
+        return Response.empty().status(HTTPStatus.OK);
     }
 }