+++ core-ng/src/main/java/core/framework/impl/web/response/BodyHandler.java
@@ -1,6 +1,5 @@
 package core.framework.impl.web.response;
 
-import core.framework.api.web.ResponseImpl;
 import core.framework.impl.web.request.RequestImpl;
 import io.undertow.io.Sender;