+++ core-ng/src/main/java/core/framework/impl/web/response/FileBodyResponseHandler.java
@@ -1,10 +1,11 @@
 package core.framework.impl.web.response;
 
-import core.framework.api.web.ResponseImpl;
 import core.framework.impl.web.request.RequestImpl;
 import io.undertow.io.IoCallback;
 import io.undertow.io.Sender;
 import io.undertow.server.HttpServerExchange;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.xnio.IoUtils;
 
 import java.io.IOException;
@@ -17,9 +18,12 @@ import java.nio.file.StandardOpenOption;
  * @author neo
  */
 class FileBodyResponseHandler implements BodyHandler {
+    private final Logger logger = LoggerFactory.getLogger(FileBodyResponseHandler.class);
+
     @Override
     public void handle(ResponseImpl response, Sender sender, RequestImpl request) {
         Path path = ((FileBody) response.body).path;
+        logger.debug("[response] file={}", path);
         try {
             FileChannel channel = FileChannel.open(path, StandardOpenOption.READ);
             sender.transferFrom(channel, new FileBodyCallback(channel));