+++ core-ng/src/main/java/core/framework/impl/web/response/TemplateBodyResponseHandler.java
@@ -1,6 +1,5 @@
 package core.framework.impl.web.response;
 
-import core.framework.api.web.ResponseImpl;
 import core.framework.impl.web.request.RequestImpl;
 import core.framework.impl.web.site.TemplateManager;
 import io.undertow.io.Sender;