+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * mongo: tweak mongo encoder and decoder code gen   
 * http: update undertow to 1.4.20, tweak the cookies config
 * search: update es to 5.6 
+* web: simplified Response interface, use chained method to set content-type and status going forward
 
 ### 4.16.2 (9/1/2017 - 9/4/2017)
 * bug: fix oracle pagination query param value